services/insights/requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
